travis: overall improvements

+++ scripts/push.sh
@@ -5,5 +5,7 @@
 # REPOSITORY
 # VERSION
 
-docker push "$REPOSITORY:$VERSION-centos-7-x86_64"
-docker rmi "$REPOSITORY:$VERSION-centos-7-x86_64"
+if [[ "$(docker images -q $REPOSITORY:$VERSION-centos-7-x86_64 2> /dev/null)" == "" ]]; then
+    docker push "$REPOSITORY:$VERSION-centos-7-x86_64"
+    docker rmi "$REPOSITORY:$VERSION-centos-7-x86_64"
+fi
